cli: reorg and renaming
