+++ cli/src/test/kotlin/ch/kleis/lcaac/cli/csv/AssessAssessCsvResultWriterTest.kt
@@ -1,5 +1,7 @@
 package ch.kleis.lcaac.cli.csv
 
+import ch.kleis.lcaac.cli.csv.assess.AssessCsvResult
+import ch.kleis.lcaac.cli.csv.assess.AssessCsvResultWriter
 import ch.kleis.lcaac.core.lang.dimension.Dimension
 import ch.kleis.lcaac.core.lang.dimension.UnitSymbol
 import ch.kleis.lcaac.core.lang.value.*
@@ -8,7 +10,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 
-class CsvResultWriterTest {
+class AssessAssessCsvResultWriterTest {
     @Test
     fun header() {
         // given
@@ -35,8 +37,8 @@ class CsvResultWriterTest {
             IndicatorValue("co2", kg) as MatrixColumnIndex<BasicNumber>
                 to QuantityValue(BasicNumber(0.5), kg)
         )
-        val result = CsvResult(request, output, impacts)
-        val writer = CsvResultWriter()
+        val result = AssessCsvResult(request, output, impacts)
+        val writer = AssessCsvResultWriter()
 
         // when
         val actual = writer.header(result)
@@ -71,8 +73,8 @@ class CsvResultWriterTest {
             IndicatorValue("co2", kg) as MatrixColumnIndex<BasicNumber>
                 to QuantityValue(BasicNumber(0.5), kg)
         )
-        val result = CsvResult(request, output, impacts)
-        val writer = CsvResultWriter()
+        val result = AssessCsvResult(request, output, impacts)
+        val writer = AssessCsvResultWriter()
 
         // when
         val actual = writer.row(result)
